Multi_Agent_System/requirements.txt
git+https://github.com/google-gemini/labs-adk.git
google-generativeai
python-dotenv
requests